Fix caught fruit radius being incorrect due to moved scale
