+++ osu.Game.Rulesets.Catch/Objects/Drawable/DrawableCatchHitObject.cs
@@ -2,11 +2,15 @@
 // See the LICENCE file in the repository root for full licence text.
 
 using System;
+using System.Collections.Generic;
+using osu.Framework.Allocation;
 using osu.Framework.Graphics;
+using osu.Framework.Graphics.Containers;
 using osu.Game.Rulesets.Objects;
 using osu.Game.Rulesets.Objects.Drawables;
 using osu.Game.Rulesets.Scoring;
 using osuTK;
+using osuTK.Graphics;
 
 namespace osu.Game.Rulesets.Catch.Objects.Drawable
 {
@@ -15,12 +19,36 @@ namespace osu.Game.Rulesets.Catch.Objects.Drawable
     {
         public override bool CanBePlated => true;
 
+        protected Container ScaleContainer;
+
         protected PalpableCatchHitObject(TObject hitObject)
             : base(hitObject)
         {
             Origin = Anchor.Centre;
             Size = new Vector2(CatchHitObject.OBJECT_RADIUS * 2);
             Masking = false;
+        }
+
+        [BackgroundDependencyLoader]
+        private void load()
+        {
+            AddRangeInternal(new Framework.Graphics.Drawable[]
+            {
+                ScaleContainer = new Container
+                {
+                    RelativeSizeAxes = Axes.Both,
+                    Origin = Anchor.Centre,
+                    Anchor = Anchor.Centre,
+                }
+            });
+
+            ScaleContainer.Scale = new Vector2(HitObject.Scale);
+        }
+
+        protected override void UpdateComboColour(Color4 proposedColour, IReadOnlyList<Color4> comboColours)
+        {
+            // ignore the incoming combo colour as we use a custom lookup
+            AccentColour.Value = comboColours[(HitObject.IndexInBeatmap + 1) % comboColours.Count];
         }
     }
 
@@ -42,6 +70,8 @@ namespace osu.Game.Rulesets.Catch.Objects.Drawable
         public virtual bool CanBePlated => false;
 
         public virtual bool StaysOnPlate => CanBePlated;
+
+        public float DisplayRadius => DrawSize.X / 2 * Scale.X * HitObject.Scale;
 
         protected DrawableCatchHitObject(CatchHitObject hitObject)
             : base(hitObject)
