+++ osu.Game.Rulesets.Catch/Objects/Drawable/DrawableDroplet.cs
@@ -1,13 +1,9 @@
 ﻿// Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
 // See the LICENCE file in the repository root for full licence text.
 
-using System.Collections.Generic;
 using osu.Framework.Allocation;
-using osu.Framework.Graphics;
-using osu.Framework.Graphics.Containers;
 using osu.Game.Rulesets.Catch.Objects.Drawable.Pieces;
 using osu.Game.Skinning;
-using osuTK;
 using osuTK.Graphics;
 
 namespace osu.Game.Rulesets.Catch.Objects.Drawable
@@ -15,8 +11,6 @@ namespace osu.Game.Rulesets.Catch.Objects.Drawable
     public class DrawableDroplet : PalpableCatchHitObject<Droplet>
     {
         public override bool StaysOnPlate => false;
-
-        protected Container ScaleContainer;
 
         public DrawableDroplet(Droplet h)
             : base(h)
@@ -26,32 +20,12 @@ namespace osu.Game.Rulesets.Catch.Objects.Drawable
         [BackgroundDependencyLoader]
         private void load()
         {
-            AddRangeInternal(new Framework.Graphics.Drawable[]
-            {
-                ScaleContainer = new Container
+            ScaleContainer.Child = new SkinnableDrawable(
+                new CatchSkinComponent(CatchSkinComponents.Droplet), _ => new Pulp
                 {
-                    RelativeSizeAxes = Axes.Both,
-                    Origin = Anchor.Centre,
-                    Anchor = Anchor.Centre,
-                    Children = new Framework.Graphics.Drawable[]
-                    {
-                        new SkinnableDrawable(
-                            new CatchSkinComponent(CatchSkinComponents.Droplet), _ => new Pulp
-                            {
-                                Size = Size / 4,
-                                AccentColour = { Value = Color4.White }
-                            })
-                    }
-                }
-            });
-
-            ScaleContainer.Scale = new Vector2(HitObject.Scale);
-        }
-
-        protected override void UpdateComboColour(Color4 proposedColour, IReadOnlyList<Color4> comboColours)
-        {
-            // ignore the incoming combo colour as we use a custom lookup
-            AccentColour.Value = comboColours[(HitObject.IndexInBeatmap + 1) % comboColours.Count];
+                    Size = Size / 4,
+                    AccentColour = { Value = Color4.White }
+                });
         }
     }
 }
