+++ osu.Game.Rulesets.Catch/Objects/Drawable/DrawableFruit.cs
@@ -2,21 +2,14 @@
 // See the LICENCE file in the repository root for full licence text.
 
 using System;
-using System.Collections.Generic;
 using osu.Framework.Allocation;
-using osu.Framework.Graphics;
-using osu.Framework.Graphics.Containers;
 using osu.Framework.Utils;
 using osu.Game.Skinning;
-using osuTK;
-using osuTK.Graphics;
 
 namespace osu.Game.Rulesets.Catch.Objects.Drawable
 {
     public class DrawableFruit : PalpableCatchHitObject<Fruit>
     {
-        private Container scaleContainer;
-
         public DrawableFruit(Fruit h)
             : base(h)
         {
@@ -26,28 +19,8 @@ namespace osu.Game.Rulesets.Catch.Objects.Drawable
         [BackgroundDependencyLoader]
         private void load()
         {
-            AddRangeInternal(new Framework.Graphics.Drawable[]
-            {
-                scaleContainer = new Container
-                {
-                    RelativeSizeAxes = Axes.Both,
-                    Origin = Anchor.Centre,
-                    Anchor = Anchor.Centre,
-                    Children = new Framework.Graphics.Drawable[]
-                    {
-                        new SkinnableDrawable(
-                            new CatchSkinComponent(getComponent(HitObject.VisualRepresentation)), _ => new FruitPiece())
-                    }
-                }
-            });
-
-            scaleContainer.Scale = new Vector2(HitObject.Scale);
-        }
-
-        protected override void UpdateComboColour(Color4 proposedColour, IReadOnlyList<Color4> comboColours)
-        {
-            // ignore the incoming combo colour as we use a custom lookup
-            AccentColour.Value = comboColours[(HitObject.IndexInBeatmap + 1) % comboColours.Count];
+            ScaleContainer.Child = new SkinnableDrawable(
+                new CatchSkinComponent(getComponent(HitObject.VisualRepresentation)), _ => new FruitPiece());
         }
 
         private CatchSkinComponents getComponent(FruitVisualRepresentation hitObjectVisualRepresentation)
